full support for rootless container backends
This mega commit introduces the concept of an autograde backend.
The "local" backend is the default backend and provides all autograde
functionality, i.e. all the sub commands that are exposed via the
command line interface.
Additionaly, the Docker/Podman backends map autograde subcommand to
autograde instances running in an isolated container environment.